refactor: Share one retained-artifact fixture across the binding tests
The edited source, the compiled target assembly, the retained artifact assembly
and the worker inputs that reference them were private to the retained-
declaration tests. The propagation tests need the same world, and two copies of
a fixture that reads its fingerprint back from a real planning run would drift.

Verification: the retained-declaration test class, 5 passed / 0 failed, single
flight.
