+++ Assets/Tests/Editor/HotReload/HotReloadRetainedArtifactFixture.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 08b475abbdbd24b309932a2d20d2fcb4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
